apps/api/requirements.txt
Flask==3.0.3
pymongo==4.10.1
python-dotenv==1.0.1
PyJWT==2.9.0
Flask-CORS==4.0.0
zai-sdk==0.0.4
requests==2.32.5
qiniu==7.12.0
